feature: early bailout if there's nothing to do

+++ bonkbot/bot.py
@@ -446,6 +446,10 @@ class BonkBot(discord.Client):
         self.__logger.debug(f"Guilds that client is part of: {guilds_discord}")
         self.__logger.debug(f"Guilds that are cached in database: {list(guilds_cached)}")
         self.__logger.debug(f"Difference: {stale_guild_ids}")
+        
+        if len(stale_guild_ids) == 0:
+            self.__logger.info(f"No stale guilds found, exiting job")
+            return
         self.__logger.info(f"Removing stale guilds from database: {stale_guild_ids}")
         self.__data_service.delete_guilds(stale_guild_ids)
